pyproject.toml
[tool.poetry]
name = "pdfp"
version = "0.0.2"
description = "PDF Processor - a GUI for some common PDF operations."
authors = ["Dreary <drearywillowdev@gmail.com>"]
license = "AGPL-3.0-only"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pyperclip = "^1.5.0"
PySide6 = "^6.5.0"
pymupdf = "^1.24.7"
gtts = "^2.5.1"
fpdf2 = "^2.7.9"

[tool.poetry.scripts]
pdfp = "pdfp.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
